ci(tests): skip apollo server on nodejs 12.x & changed verification for deepkit
