+++ test/frameworks/http-deepkit.framework.spec.ts
@@ -1,6 +1,6 @@
 describe('HttpDeepkitFramework', () => {
   // very ugly, i know, but i didn't find any other way, sorry
-  if (process.env.SKIP_DEEPKIT === 'true') {
+  if (process.env.TEST_NODE_VERSION === '12.x') {
     it('when nodejs is 12.x', () => {
       expect(true).toEqual(true);
     });
